+++ src/components/collections/Form/component.js
@@ -8,7 +8,7 @@ import { Heading } from 'typography/Heading';
 import { Link } from 'elements/Link';
 
 import { getValidationWithDefaults } from './utils/getValidationWithDefaults';
-import { getIsRequiredError } from './utils/getIsRequiredError';
+import { getIsRequiredErrorThenSetState } from './utils/getIsRequiredErrorThenSetState';
 import { getIsValidError } from './utils/getIsValidError';
 import { getEmptyRequiredInputs } from './utils/getEmptyRequiredInputs';
 import { getFormChild } from './utils/getFormChild';
@@ -21,15 +21,26 @@ import { getIsSubmitButtonDisabled } from './utils/getIsSubmitButtonDisabled';
 export class Component extends PureComponent {
   state = {};
 
-  handleInputBlur = name => {
-    const validation = getValidationWithDefaults(this.props.validation[name]);
-    const hasError = getIsRequiredError(validation, this.state[name]);
-
-    hasError &&
-      this.setState({ [name]: { error: validation.isRequiredMessage } });
-  };
+  handleInputBlur = name =>
+    getIsRequiredErrorThenSetState(
+      this.setState.bind(this),
+      this.props.validation[name],
+      name,
+      this.state[name]
+    );
 
   handleInputChange = (name, value) => {
+    const hasIsRequiredError = getIsRequiredErrorThenSetState(
+      this.setState.bind(this),
+      this.props.validation[name],
+      name,
+      { value }
+    );
+
+    if (hasIsRequiredError) {
+      return;
+    }
+
     const { invalidMessage, getIsValid } = getValidationWithDefaults(
       this.props.validation[name]
     );
@@ -59,6 +70,7 @@ export class Component extends PureComponent {
   render = () => {
     const {
       actionLink,
+      autoComplete,
       children,
       errorMessage,
       headingText,
@@ -75,6 +87,7 @@ export class Component extends PureComponent {
         )}
         <Card.Content>
           <Form
+            autoComplete={autoComplete}
             error={!!errorMessage}
             onSubmit={this.handleSubmit}
             success={!!successMessage}
@@ -103,6 +116,7 @@ export class Component extends PureComponent {
 Component.displayName = 'Form';
 
 Component.defaultProps = {
+  autoComplete: null,
   errorMessage: '',
   headingText: null,
   onSubmit: Function.prototype,
@@ -120,6 +134,8 @@ Component.propTypes = {
     /** The visible text for the secondary call to action */
     text: PropTypes.node.isRequired,
   }),
+  /** Can inputs be completed automatically by the browser. See [MDN docs for `<form />` for more](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/form). */
+  autoComplete: PropTypes.oneOf(['on', 'off']),
   /** The child components and elements. */
   children: PropTypes.node.isRequired,
   /** The message to display when the form has an error */