+++ src/components/general-widgets/Contact/component.js
@@ -51,6 +51,7 @@ export const Component = ({
   validation,
 }) => (
   <Form
+    autoComplete="off"
     errorMessage={errorMessage}
     headingText={headingText}
     onSubmit={onSubmit}
@@ -59,10 +60,10 @@ export const Component = ({
     validation={validation}
   >
     <InputGroup>
-      <TextInput label={nameInputLabel} name="name" />
-      <PhoneInput label={phoneInputLabel} name="phone" />
+      <TextInput autoComplete="name" label={nameInputLabel} name="name" />
+      <PhoneInput autoComplete="tel" label={phoneInputLabel} name="phone" />
     </InputGroup>
-    <TextInput label={emailInputLabel} name="email" />
+    <TextInput autoComplete="email" label={emailInputLabel} name="email" />
     <InputGroup>
       <DateRangePicker
         endDatePlaceholderText={departureDateInputLabel}