+++ src/components/general-widgets/FeaturedProperty/component.js
@@ -17,6 +17,8 @@ export const Component = ({
   bedroomsNumber,
   guestsNumber,
   imageAlternativeText,
+  imageSizes,
+  imageSrcSet,
   imageUrl,
   locationName,
   nightPrice,
@@ -32,6 +34,8 @@ export const Component = ({
       imageUrl={imageUrl}
       isFluid
       placeholderImageUrl={placeholderImageUrl}
+      sizes={imageSizes}
+      srcSet={imageSrcSet}
     />
     <Card.Content>
       <Card.Meta>
@@ -54,7 +58,9 @@ Component.displayName = 'FeaturedProperty';
 
 Component.defaultProps = {
   bedroomsNumber: null,
-  imageAlternativeText: '',
+  imageAlternativeText: undefined,
+  imageSizes: undefined,
+  imageSrcSet: undefined,
   placeholderImageUrl: undefined,
 };
 
@@ -65,6 +71,10 @@ Component.propTypes = {
   guestsNumber: PropTypes.number.isRequired,
   /** The alternative text for the image to display. */
   imageAlternativeText: PropTypes.string,
+  /** A list of one or more strings separated by commas indicating a set of source sizes. See [the MDN docs for more information](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/img). */
+  imageSizes: PropTypes.string,
+  /** A list of one or more strings separated by commas indicating a set of possible image sources for the user agent to use. See [the MDN docs for more information](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/img). */
+  imageSrcSet: PropTypes.string,
   /** URL pointing to the image to display. */
   imageUrl: PropTypes.string.isRequired,
   /** The name of the location of the property. */