+++ src/components/general-widgets/FeaturedRoomTypes/component.spec.js
@@ -1,17 +1,6 @@
 import React from 'react';
-import { shallow } from 'enzyme';
-import {
-  expectComponentToBe,
-  expectComponentToHaveChildren,
-  expectComponentToHaveProps,
-  expectComponentToHaveDisplayName,
-} from '@lodgify/enzyme-jest-expect-helpers';
-
-import { getArrayOfLengthOfItem } from 'utils/get-array-of-length-of-item';
-import { Grid } from 'layout/Grid';
-import { GridColumn } from 'layout/GridColumn';
-import { FeaturedRoomType } from 'general-widgets/FeaturedRoomType';
-import { Heading } from 'typography/Heading';
+import { mount } from 'enzyme';
+import { expectComponentToHaveDisplayName } from '@lodgify/enzyme-jest-expect-helpers';
 
 import { Component as FeaturedRoomTypes } from './component';
 
@@ -41,94 +30,25 @@ const featuredRoomTypes = [
 ];
 
 const getFeaturedRoomTypes = otherProps =>
-  shallow(
+  mount(
     <FeaturedRoomTypes featuredRoomTypes={featuredRoomTypes} {...otherProps} />
   );
 
 describe('<FeaturedRoomTypes />', () => {
-  it('should have `Grid` component as a wrapper', () => {
-    const wrapper = getFeaturedRoomTypes();
+  describe('by default', () => {
+    it('should render the right structure', () => {
+      const actual = getFeaturedRoomTypes();
 
-    expectComponentToBe(wrapper, Grid);
-  });
-
-  describe('if `props.headingText` is not passed', () => {
-    describe('the `Grid` component', () => {
-      it('should render the right children', () => {
-        const wrapper = getFeaturedRoomTypes().find(Grid);
-
-        expectComponentToHaveChildren(
-          wrapper,
-          ...getArrayOfLengthOfItem(2, GridColumn)
-        );
-      });
-    });
-
-    describe('each of the array of  `GridColumns`', () => {
-      const getFirstGridColumn = () =>
-        getFeaturedRoomTypes()
-          .find(GridColumn)
-          .first();
-
-      it('should have the right right props', () => {
-        const wrapper = getFirstGridColumn();
-
-        expectComponentToHaveProps(wrapper, {
-          computer: 4,
-          mobile: 12,
-          tablet: 6,
-        });
-      });
-
-      it('should render the right children', () => {
-        const wrapper = getFirstGridColumn();
-
-        expectComponentToHaveChildren(wrapper, FeaturedRoomType);
-      });
+      expect(actual).toMatchSnapshot();
     });
   });
 
   describe('if `props.headingText` is passed', () => {
-    const headingText = 'My god I love Elton John';
+    it('should render the right structure', () => {
+      const headingText = 'My god I love Elton John';
+      const actual = getFeaturedRoomTypes({ headingText });
 
-    describe('the `Grid` component', () => {
-      it('should render the right children', () => {
-        const wrapper = getFeaturedRoomTypes({ headingText }).find(Grid);
-
-        expectComponentToHaveChildren(
-          wrapper,
-          ...getArrayOfLengthOfItem(3, GridColumn)
-        );
-      });
-    });
-
-    describe('the heading `GridColumn` component', () => {
-      const getFirstGridColumn = () =>
-        getFeaturedRoomTypes({ headingText })
-          .find(GridColumn)
-          .first();
-
-      it('should have the right right props', () => {
-        const wrapper = getFirstGridColumn();
-
-        expectComponentToHaveProps(wrapper, {
-          width: 12,
-        });
-      });
-
-      it('should render the right children', () => {
-        const wrapper = getFirstGridColumn();
-
-        expectComponentToHaveChildren(wrapper, Heading);
-      });
-    });
-
-    describe('the `Heading` component', () => {
-      it('should render the right children', () => {
-        const wrapper = getFeaturedRoomTypes({ headingText }).find(Heading);
-
-        expectComponentToHaveChildren(wrapper, headingText);
-      });
+      expect(actual).toMatchSnapshot();
     });
   });