+++ src/components/inputs/CaptchaInput/component.js
@@ -23,6 +23,7 @@ export const Component = ({
     </Form.Field>
     <Form.Field>
       <TextInput
+        autoComplete="off"
         error={error}
         isValid={isValid}
         label={label}