+++ src/components/inputs/Dropdown/utils/getHasImages.spec.js
@@ -1,9 +1,9 @@
 import { getHasImages } from './getHasImages';
 
 describe('getHasImages', () => {
-  it('should return `true` if object has `image` property', () => {
+  it('should return `true` if object has `imageUrl` property', () => {
     const options = [
-      { value: 'firstValue', image: '/yoyo/' },
+      { value: 'firstValue', imageUrl: '/yoyo/' },
       { value: 'secondValue' },
     ];
     const actual = getHasImages(options);
@@ -11,8 +11,38 @@ describe('getHasImages', () => {
     expect(actual).toBe(true);
   });
 
-  it('should return `false` if object doesnt have `image` property', () => {
+  it('should return `true` if object has `imageSizes` and `imageSrcSet` property', () => {
+    const options = [
+      { value: 'firstValue', imageSizes: 'sizes', imageSrcSet: 'srcSet' },
+      { value: 'secondValue' },
+    ];
+    const actual = getHasImages(options);
+
+    expect(actual).toBe(true);
+  });
+
+  it('should return `false` if object doesnt have `imageUrl` property', () => {
     const options = [{ value: 'firstYoValue' }, { value: 'secondYoValue' }];
+    const actual = getHasImages(options);
+
+    expect(actual).toBe(false);
+  });
+
+  it('should return `false` if object only has `imageSizes` property', () => {
+    const options = [
+      { value: 'firstYoValue', imageSizes: 'sizes' },
+      { value: 'secondYoValue' },
+    ];
+    const actual = getHasImages(options);
+
+    expect(actual).toBe(false);
+  });
+
+  it('should return `false` if object only has `imageSrcSet` property', () => {
+    const options = [
+      { value: 'firstYoValue', imageSrcSet: 'srcSet' },
+      { value: 'secondYoValue' },
+    ];
     const actual = getHasImages(options);
 
     expect(actual).toBe(false);