+++ src/components/property-page-widgets/PropertyPageSearchBar/component.spec.js
@@ -13,6 +13,7 @@ import { Component as PropertyPageSearchBar } from './component';
 const { guestsOptions } = require('./mock-data/options');
 
 const props = {
+  getIsDayBlocked: Function.prototype,
   guestsOptions: guestsOptions,
   isFixed: true,
   modalTrigger: (
@@ -21,6 +22,7 @@ const props = {
     </Button>
   ),
   onChangeInput: Function.prototype,
+  onSubmit: Function.prototype,
   searchButton: (
     <Button isPositionedRight isRounded>
       {CHECK_OUR_AVAILABILITY}